refactor: lift OperationAmount to widgets; widen operation transformers
- Move OperationAmount from features/multisig-operations into a new
  widgets/transaction-amount module so drafts and the dashboard queue stop
  cross-feature-importing it.
- Widen the title and icon transformer signatures to a minimal
  { transaction, chainId } context. Non-multisig callers (drafts) can now
  feed real data through the transformer pipeline without 'as never' casts.
