+++ src/renderer/features/drafts/components/DraftRow.tsx
@@ -14,9 +14,9 @@ import { contactModel } from '@/entities/contact';
 import { networkModel } from '@/entities/network';
 import { accountUtils, walletModel } from '@/entities/wallet';
 import { authModel } from '@/aggregates/backend';
-import { OperationAmount } from '@/features/multisig-operations';
 import { WalletPairingOperationTrigger } from '@/features/wallet-pairing';
 import { NamedAccount } from '@/widgets/NameResolver';
+import { OperationAmount } from '@/widgets/transaction-amount';
 import { draftAccountsOverviewSlot } from '../lib/draft-row-slot';
 import { useDraftOperationTitle } from '../lib/useDraftOperationTitle';
 import { useDraftTransactionAmount } from '../lib/useDraftTransactionAmount';
