+++ src/renderer/features/multisig-operations/index.ts
@@ -4,7 +4,6 @@ export { operationDetailsSlot } from './components/OperationFullInfo';
 export { operationOverviewSlot } from './components/OperationSignatories';
 export { OperationIcon, operationIconTransformer } from './components/OperationIcon';
 export { OperationActions } from './components/OperationActions';
-export { OperationAmount } from './components/OperationAmount';
 export { confirmTransactionInfoSlot } from './components/ActionSteps/Confirmation';
 export { Operations } from './components/Operations';
 export { Search } from './components/Search';
