+++ src/renderer/widgets/transaction-amount/OperationAmount.tsx
