+++ src/renderer/widgets/transaction-amount/index.ts
@@ -1,0 +1,1 @@
+export { OperationAmount } from './OperationAmount';
